+++ src/api_routes.php
@@ -4,11 +4,13 @@
 
         Route::post('addRole', "{$namespaceController}\LaraminRoleController@store");
         Route::put('editRole', "{$namespaceController}\LaraminRoleController@update");
-        Route::delete('deleteRole/{id}', "{$namespaceController}\LaraminRoleController@destroy");
+        Route::delete('deleteRole/{auth}/{id}', "{$namespaceController}\LaraminRoleController@destroy");
+
         Route::put('editPermissionRole',"{$namespaceController}\LaraminRoleController@assignPermission");
         Route::get('getPermissionRole',"{$namespaceController}\LaraminRoleController@getAssignPermission");
 
+        Route::post('addUser', "{$namespaceController}\LaraminUserController@store");
         Route::put('editUser',"{$namespaceController}\LaraminUserController@update");
-        Route::delete('deleteUser/{id}',"{$namespaceController}\LaraminUserController@destroy");
+        Route::delete('deleteUser/{auth}/{id}',"{$namespaceController}\LaraminUserController@destroy");
 
     });
